phase 4 of issue #159
