+++ docs/api/geometry.md
@@ -10,9 +10,10 @@ Rust, and covers the rings directly. A backend is still required for WKT ingest
 a backend geometry object by definition.
 
 The spherical outline machinery behind `to_geometry(dissolve=True)` lives in
-`mortie.dissolve` (issue #159), mirroring `src_rust/src/dissolve.rs`. It has no
-public members, so it has no page of its own; `to_geometry` is still where the
-whole emit path is documented.
+`mortie.dissolve` (issue #159), mirroring `src_rust/src/dissolve.rs`, and the
+backend gate plus the codec quartet live in `mortie.codec`. Neither has any
+public member, so neither has a page of its own; the functions below are still
+where the whole ingest/emit path is documented.
 
 ::: mortie.geometry
     options:
